+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.13.4] - 2026-03-11
+
+### Changed
+- (Admin) Removed Users tab from admin panel to clean up space
+
+### Fixed
+- Logging in from the admin page now returns you to admin instead of the home page
+
 ## [0.13.3] - 2026-03-11
 
 ### Added